requirements.txt
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

pandas>=1.5
numpy>=1.23
scipy>=1.9

PyYAML>=6.0
joblib>=1.3

matplotlib>=3.7
seaborn>=0.12
